requirements.txt
ace-tools
faker
pandas
fastapi
uvicorn
python-dotenv
ollama
